entity id를 inner class로 변경 Many to one fetch변경
