+++ src/main/java/gift/repository/OptionRepository.java
@@ -1,7 +1,6 @@
 package gift.repository;
 
 import gift.dto.product.ProductWithOptionDTO;
-import gift.entity.compositeKey.OptionId;
 import gift.entity.Option;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface OptionRepository extends JpaRepository<Option, OptionId> {
+public interface OptionRepository extends JpaRepository<Option, Option.OptionId> {
     @Query("delete from Option u where u.id = :product_id")
     void deleteByProductID(@Param("product_id") int id);
