+++ src/main/java/gift/repository/WishListRepository.java
@@ -1,7 +1,6 @@
 package gift.repository;
 
 import gift.dto.product.ShowProductDTO;
-import gift.entity.compositeKey.WishListId;
 import gift.entity.WishList;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -13,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface WishListRepository extends JpaRepository<WishList, WishListId> {
+public interface WishListRepository extends JpaRepository<WishList, WishList.WishListId> {
     @Query("SELECT new gift.dto.product.ShowProductDTO(p.id, p.name, p.price, p.imageUrl) " +
             "FROM Product p join WishList w  on p.id = w.id.productId where w.id.userId = :userId")
     Page<ShowProductDTO> findByUserId(@Param("userId") int tokenUserId, Pageable pageable);
